+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.1-dev
+
+Added `htmlEntryPoint` option to the transformer. This will search that file for
+any script tag whose src is `entryPoint` and rewrite it to point at the
+bootstrapped file `newEntryPoint`.
+
 ## 0.1.0+1
 
 Quick fix for the transformer on windows.